requirements.txt
attrs==25.4.0
audioread==3.0.1
babel==2.17.0
brotli==1.2.0
certifi==2025.4.26
cffi==1.15.1
charset-normalizer==3.4.2
click==8.1.8
cssselect2==0.8.0
csvw==3.7.0
cycler==0.11.0
decorator==5.1.1
dlinfo==2.0.0
filelock==3.20.2
Flask==2.2.5
fonttools==4.61.1
fsspec==2025.12.0
idna==3.10
importlib-metadata==6.7.0
isodate==0.7.2
itsdangerous==2.1.2
Jinja2==3.1.6
jiwer==3.0.4
joblib==1.3.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.5
language-tags==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.39.1
MarkupSafe==2.1.5
matplotlib==3.5.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.4.2
noisereduce==3.0.3
numba==0.56.4
numpy==1.21.6
openai-whisper==20231117
packaging==24.0
pandas==1.3.5
phonemizer==3.2.1
Pillow==9.5.0
platformdirs==4.0.0
pooch==1.8.2
praat-parselmouth==0.4.5
pycparser==2.21
pydyf==0.12.1
pyparsing==3.1.4
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
RapidFuzz==3.14.3
rdflib==7.5.0
referencing==0.37.0
regex==2025.11.3
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.30.0
scikit-learn==1.0.2
scipy==1.7.3
segments==2.3.0
six==1.17.0
soundfile==0.13.1
soxr==0.3.7
sympy==1.14.0
termcolor==3.3.0
threadpoolctl==3.1.0
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
torch==2.1.2
tqdm==4.67.1
typing_extensions==4.7.1
uritemplate==4.2.0
urllib3==2.0.7
waitress==2.1.2
weasyprint==67.0
webencodings==0.5.1
Werkzeug==2.2.3
zipp==3.15.0
zopfli==0.4.0
